refactor: isolate hosted operator ui helpers
